Feat adn Refactor: Refactor the add course page in the profile
The add course page in the profile was refactored to improve the user experience. Changes were made to the placement of some elements, and new features were added, such as the option to attach a course image and validate the expedition date. Additionally, a confirmation pop-up was implemented when saving the course information.